requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==4.2.10
django-cors-headers==4.3.1
djangorestframework==3.14.0
exceptiongroup==1.2.2
idna==3.10
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
pytest==8.3.5
pytest-django==4.10.0
pytz==2025.1
requests==2.32.3
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.12.2
urllib3==2.3.0
